% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areametric.R
\name{areametric}
\alias{areametric}
\title{Areametric Down-scaling}
\usage{
areametric(buildings, blocks, code, pop, area)
}
\arguments{
\item{buildings}{object of class \code{sf}}

\item{blocks}{object of class \code{sf}}

\item{code}{string of block id column}

\item{pop}{string of block pop column}

\item{area}{string of building area column}
}
\value{
an object of class \code{sf} including population counts
}
\description{
Population down-scaling from a coarse source zone to a finer scale using the areametric approach
}
\examples{
    library(populR)
    data("buildings")
    data("blocks")
    population_a = areametric(buildings = buildings, blocks = blocks,
        code = 'code', pop = 'pop', area = 'area')

}
\references{
Lwin, K. K., & Murayama, Y. (2009)
    \emph{A GIS approach to estimation of building population for micro-spatial analysis. Transactions in GIS, 13(4), 401–414.}
    \doi{https://doi.org/10.1111/j.1467-9671.2009.01171.x}
}
\seealso{
\link[populR]{volumetric}
}
