% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\docType{data}
\name{blocks}
\alias{blocks}
\title{Blocks}
\format{
object of \code{sf} class with 6 rows and 3 columns:
\describe{
  \item{\code{code}}{block id}
  \item{\code{pop}}{population count}
  \item{\code{geometry}}{geometry}
}
}
\source{
\url{http://www.mbatsaris.gr/}
}
\usage{
blocks
}
\description{
object of \code{sf} class representing the blocks of a fictional area
}
\keyword{datasets}
