% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\docType{data}
\name{buildings}
\alias{buildings}
\title{Buildings}
\format{
object of \code{sf} class with 19 rows and 4 columns:
\describe{
  \item{\code{area}}{projected area}
  \item{\code{floors}}{number of floors}
  \item{\code{bcode}}{building id}
  \item{\code{geometry}}{geometry}
}
}
\source{
\url{http://mbatsaris.gr/}
}
\usage{
buildings
}
\description{
object of \code{sf} class representing the buildings of a fictional areag
}
\keyword{datasets}
