% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_a.R
\name{pp_a}
\alias{pp_a}
\title{Calculate Area/Volume for Target Zone Features}
\usage{
pp_a(target, volume = NULL)
}
\arguments{
\item{target}{object of class \code{sf}}

\item{volume}{(optional) target number of floors or height values of target features}
}
\value{
an object of class \code{sf} including area or volume values
}
\description{
Calculate Area/Volume for Target Zone Features
}
\examples{
    library(populR)
    data("target")
    data("source")

    # area calculation
    pp_a(target = target)

    # volume calculation
    pp_a(target = target, volume = floors)

}
