% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{RMS Error}
\usage{
rmse(buildings, blocks, blockcode, blockpop, buildpop, title)
}
\arguments{
\item{buildings}{buildings file in \code{sf} format}

\item{blocks}{census blocks file in \code{sf} format}

\item{blockcode}{block id column \code{string}}

\item{blockpop}{block population column name \code{string}}

\item{buildpop}{building population column name \code{string}}

\item{title}{title of scatterplot \code{string}}
}
\value{
rms error, correlation and scatterplot
}
\description{
This function calculates the rmse between the source and target counts
}
