% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\docType{data}
\name{source}
\alias{source}
\title{Source}
\format{
object of \code{sf} class with 6 rows and 3 columns:
\describe{
  \item{\code{id}}{source id}
  \item{\code{pop}}{population count}
  \item{\code{geometry}}{geometry}
}
}
\source{
\url{http://www.mbatsaris.gr/}
}
\usage{
source
}
\description{
object of \code{sf} class representing the blocks of a fictional area
}
\keyword{datasets}
