% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\docType{data}
\name{src}
\alias{src}
\title{Source (src)}
\format{
object of \code{sf} class with 9 rows and 3 columns:
\describe{
  \item{\code{sid}}{Source identification number}
  \item{\code{pop}}{Source population values to be interpolated}
  \item{\code{geometry}}{Geometry}
}
}
\source{
\url{http://www.mbatsaris.gr/}
}
\usage{
src
}
\description{
object of \code{sf} class representing the blocks of a fictional area
}
\keyword{datasets}
