% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\docType{data}
\name{target}
\alias{target}
\title{Target}
\format{
object of \code{sf} class with 19 rows and 2 columns:
\describe{
  \item{\code{floors}}{number of floors}
  \item{\code{geometry}}{geometry}
}
}
\source{
\url{http://mbatsaris.gr/}
}
\usage{
target
}
\description{
object of \code{sf} class representing the buildings of a fictional areag
}
\keyword{datasets}
