% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trg.R
\docType{data}
\name{trg}
\alias{trg}
\title{Target (trg)}
\format{
object of \code{sf} class with 179 rows and 3 columns:
\describe{
  \item{\code{tid}}{target id}
  \item{\code{floors}}{number of floors}
  \item{\code{geometry}}{geometry}
}
}
\source{
\url{http://mbatsaris.gr/}
}
\usage{
trg
}
\description{
object of \code{sf} class representing the buildings of a fictional area
}
\keyword{datasets}
